+++ actix/src/lib.rs
@@ -53,7 +53,7 @@ pub mod utils;
 
 #[cfg(feature = "macros")]
 pub use actix_derive::{main, test, Message, MessageResponse};
-pub use actix_rt::{spawn, Arbiter, System, SystemRunner};
+pub use actix_rt::{spawn, Arbiter, ArbiterHandle, System, SystemRunner};
 
 pub use crate::actor::{
     Actor, ActorContext, ActorState, AsyncContext, Running, SpawnHandle, Supervised,
@@ -88,7 +88,7 @@ pub mod prelude {
     #[cfg(feature = "macros")]
     pub use actix_derive::{Message, MessageResponse};
 
-    pub use actix_rt::{Arbiter, System, SystemRunner};
+    pub use actix_rt::{Arbiter, ArbiterHandle, System, SystemRunner};
 
     pub use crate::actor::{
         Actor, ActorContext, ActorState, AsyncContext, Running, SpawnHandle, Supervised,
@@ -136,7 +136,7 @@ pub mod dev {
         pub use crate::address::channel::{channel, AddressReceiver, AddressSender};
     }
     pub use crate::contextimpl::{AsyncContextParts, ContextFut, ContextParts};
-    pub use crate::handler::{MessageResponse, ResponseChannel};
+    pub use crate::handler::{MessageResponse, OneshotSender};
     pub use crate::mailbox::Mailbox;
     pub use crate::registry::{Registry, SystemRegistry};
 }